Initial commit
A sample movie app to demonstrate keystone.js, angular and cloudinary. A
am able to upload multiple cloudinary images using angular and send to
keystone with the exact same format expected but not using direct upload
causing keystone to discard images. Need to fix this.
Note: This issue occurs only when uploading multiple images but not
single image (aka. heroImage) not sure why.